+++ packages/bundler/src/submitter.ts
@@ -78,6 +78,10 @@ export class BundlerSubmitter {
   }
 
   async submitBatch(operations: ProvenOperation[]): Promise<void> {
+    // TODO: this job isn't idempotent. if one step fails, bullmq will re-try which may cause issues
+    // current plan is to mark reverted bundles as failed.
+    // will circle back after further testing and likely re-queue/re-validate ops in the reverted bundle"
+
     // Loop through current batch and set each job status to IN_FLIGHT
     const inflightStatusTransactions = operations.map((op) => {
       const jobId = computeOperationDigest(op).toString();
@@ -96,10 +100,37 @@ export class BundlerSubmitter {
 
     console.log("submitting bundle...");
     // Hardcode gas limit to skip eth_estimateGas
-    const tx = await this.walletContract.processBundle(
-      { operations },
-      { gasLimit: 1_000_000 }
-    );
+    // * there's gotta be a better way to handle this error (error handling logic is async)
+    let tx;
+    try {
+      tx = await this.walletContract.processBundle(
+        { operations },
+        { gasLimit: 1_000_000 }
+      );
+    } catch (err) {
+      console.log("failed to process bundle:", err);
+      const statusTxs = operations.map((op) => {
+        const digest = computeOperationDigest(op);
+        console.log(
+          `setting operation with digest ${digest} to BUNDLE_REVERTED`
+        );
+        return this.statusDB.getSetJobStatusTransaction(
+          digest.toString(),
+          OperationStatus.BUNDLE_REVERTED
+        );
+      });
+
+      await this.redis.multi(statusTxs).exec((maybeErr) => {
+        if (maybeErr) {
+          throw new Error(
+            `failed to set job status transactions after bundle reverted: ${maybeErr}`
+          );
+        }
+      });
+
+      return;
+    }
+
     console.log("waiting for confirmation...");
     const receipt = await tx.wait(1);
 
@@ -131,7 +162,7 @@ export class BundlerSubmitter {
     await this.redis.multi(executedStatusTransactions).exec((maybeErr) => {
       if (maybeErr) {
         throw new Error(
-          `failed to set job status transactions post-op: ${maybeErr}`
+          `failed to set job status transactions after bundle executed: ${maybeErr}`
         );
       }
     });